# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Food Journey API",
    "description": "Food Journey 应用的后端 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/profile": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "Get Profile",
        "description": "获取用户个人资料\n\n返回当前登录用户的详细信息\n\nArgs:\n    request: 请求对象\n    current_user: 当前登录用户\n    \nReturns:\n    JSONResponse: 用户详细信息",
        "operationId": "get_profile_api_v1_auth_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "用户档案"
        ],
        "summary": "Get Profile",
        "description": "获取用户的完整画像信息\n\nArgs:\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    CompleteProfile: 包含用户所有档案信息的响应",
        "operationId": "get_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/basic": {
      "put": {
        "tags": [
          "用户档案"
        ],
        "summary": "Update Basic Info",
        "description": "更新用户的基础信息和健康数据\n\nArgs:\n    data: 基础信息更新数据\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    UpdateResponse: 包含更新结果的响应",
        "operationId": "update_basic_info_api_v1_profile_basic_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicInfoUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/diet": {
      "put": {
        "tags": [
          "用户档案"
        ],
        "summary": "Update Diet Preferences",
        "description": "更新用户的饮食偏好\n\nArgs:\n    data: 饮食偏好更新数据\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    UpdateResponse: 包含更新结果的响应",
        "operationId": "update_diet_preferences_api_v1_profile_diet_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DietPreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/fitness": {
      "put": {
        "tags": [
          "用户档案"
        ],
        "summary": "Update Fitness Preferences",
        "description": "更新用户的健身偏好",
        "operationId": "update_fitness_preferences_api_v1_profile_fitness_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FitnessPreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/stats": {
      "get": {
        "tags": [
          "用户档案"
        ],
        "summary": "Get Health Stats",
        "description": "获取用户的综合健康数据统计\n\nArgs:\n    start_date: 开始日期（可选）\n    end_date: 结束日期（可选）\n    stat_type: 统计类型，默认为daily\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    HealthStatsResponse: 包含健康数据统计的响应",
        "operationId": "get_health_stats_api_v1_profile_stats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "stat_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "daily",
              "title": "Stat Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/exercise": {
      "post": {
        "tags": [
          "用户档案"
        ],
        "summary": "Record Exercise",
        "description": "记录用户的运动数据\n\nArgs:\n    data: 运动记录数据\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    ExerciseRecord: 创建的运动记录",
        "operationId": "record_exercise_api_v1_profile_exercise_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/meal": {
      "post": {
        "tags": [
          "用户档案"
        ],
        "summary": "Record Meal",
        "description": "记录用户的餐食数据\n\nArgs:\n    data: 餐食记录数据\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    MealRecord: 创建的餐食记录",
        "operationId": "record_meal_api_v1_profile_meal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/nutrition/summary/{date}": {
      "get": {
        "tags": [
          "用户档案"
        ],
        "summary": "Get Daily Nutrition Summary",
        "description": "获取指定日期的营养摄入汇总\n\nArgs:\n    date: 日期字符串 (YYYY-MM-DD)\n    current_user: 当前用户\n    db: 数据库会话\n\nReturns:\n    DailyNutritionSummary: 每日营养摄入汇总",
        "operationId": "get_daily_nutrition_summary_api_v1_profile_nutrition_summary__date__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyNutritionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "用户名"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "description": "头像URL"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login",
            "description": "最后登录时间"
          },
          "login_count": {
            "type": "integer",
            "title": "Login Count",
            "description": "登录次数",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "用户信息响应模型",
        "example": {
          "avatar_url": "http://example.com/avatar.jpg",
          "created_at": "2024-01-01T00:00:00",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "is_active": true,
          "last_login": "2024-01-02T10:30:00",
          "login_count": 5,
          "username": "testuser"
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}